requirements.txt
streamlit
python-dotenv
openai
chromadb
pymupdf
tiktoken
pytest
